+++ functional_tests/main.go
@@ -10,6 +10,7 @@ import (
 	"net/http"
 	"os"
 	"os/exec"
+	"path"
 	"strings"
 	"time"
 
@@ -86,16 +87,11 @@ func initializeScenario(ctx *godog.ScenarioContext) {
 	ctx.Step(`^I delete the Canary "([^"]*)"$`, iDeleteTheCanary)
 
 	ctx.After(func(ctx context.Context, sc *godog.Scenario, scenarioError error) (context.Context, error) {
-		showLog := false
-		if scenarioError != nil {
-			showLog = true
-		}
-
 		if server != nil {
 			err := server.Shutdown()
 			if err != nil {
 				logger.Error("Unable to shutdown server", "error", err)
-				showLog = true
+				scenarioError = err
 			}
 		}
 
@@ -104,17 +100,26 @@ func initializeScenario(ctx *godog.ScenarioContext) {
 			err := executeCommand([]string{"/usr/local/bin/shipyard", "destroy"}, false)
 			if err != nil {
 				logger.Error("Unable to destroy shipyard resources", "error", err)
-				showLog = true
+				scenarioError = err
 			}
 		}
 
-		if showLog && !*alwaysLog {
-			fmt.Println(logStore.String())
+		if scenarioError != nil && !*alwaysLog {
+			pwd, _ := os.Getwd()
+			logfile := path.Join(pwd, "tests.log")
+			// create log file
+			os.Remove(logfile)
+			os.WriteFile(logfile, logStore.Bytes(), os.ModePerm)
+
+			d, _ := ioutil.ReadFile(logfile)
+			fmt.Printf("%s\n", string(d))
+
+			fmt.Printf("Error log written to file %s", logfile)
 		}
 
 		// exit after the scenario when don't destroy is set
 		if *dontDestroy {
-			if showLog {
+			if scenarioError != nil {
 				os.Exit(1)
 			}
 
@@ -160,8 +165,8 @@ func startServer() error {
 func retryOperation(f func() error) error {
 	// max time to wait 300s
 	attempt := 0
-	maxAttempts := 60
-	delay := 5 * time.Second
+	maxAttempts := 100
+	delay := 10 * time.Second
 
 	var funcError error
 	for attempt = 0; attempt < maxAttempts; attempt++ {