+++ plugins/consul/plugin_test.go
@@ -25,12 +25,16 @@ func setupPlugin(t *testing.T) (*Plugin, *clients.ConsulMock) {
 	mc.On("CreateServiceDefaults", mock.Anything).Return(nil)
 	mc.On("CreateServiceResolver", mock.Anything).Return(nil)
 	mc.On("CreateServiceRouter", mock.Anything, mock.Anything).Return(nil)
+	mc.On("CreateUpstreamRouter", mock.Anything, mock.Anything).Return(nil)
 	mc.On("CreateServiceSplitter", mock.Anything, mock.Anything, mock.Anything).Return(nil)
+	mc.On("CreateServiceIntention", mock.Anything, mock.Anything).Return(nil)
 
+	mc.On("DeleteServiceSplitter", mock.Anything).Return(nil)
 	mc.On("DeleteServiceDefaults", mock.Anything).Return(nil)
 	mc.On("DeleteServiceResolver", mock.Anything).Return(nil)
 	mc.On("DeleteServiceRouter", mock.Anything).Return(nil)
-	mc.On("DeleteServiceSplitter", mock.Anything).Return(nil)
+	mc.On("DeleteUpstreamRouter", mock.Anything).Return(nil)
+	mc.On("DeleteServiceIntention", mock.Anything).Return(nil)
 
 	data := testutils.GetTestData(t, "valid_kubernetes_release.json")
 	dep := map[string]interface{}{}
@@ -71,9 +75,11 @@ func TestSetupCreatesConsulServiceDefaults(t *testing.T) {
 	require.NoError(t, err)
 
 	mc.AssertCalled(t, "CreateServiceDefaults", "api")
+	mc.AssertCalled(t, "CreateServiceDefaults", clients.ControllerServiceName)
+	mc.AssertCalled(t, "CreateServiceDefaults", clients.UpstreamRouterName)
 }
 
-func TestSetupFailsOnCreateServiceDefaultsError(t *testing.T) {
+func TestSetupFailsOnCreateServiceDefaultsForServiceError(t *testing.T) {
 	p, mc := setupPlugin(t)
 
 	testutils.ClearMockCall(&mc.Mock, "CreateServiceDefaults")
@@ -82,6 +88,41 @@ func TestSetupFailsOnCreateServiceDefaultsError(t *testing.T) {
 	err := p.Setup(context.Background())
 
 	mc.AssertCalled(t, "CreateServiceDefaults", "api")
+	mc.AssertNotCalled(t, "CreateServiceDefaults", clients.ControllerServiceName)
+	mc.AssertNotCalled(t, "CreateServiceDefaults", clients.UpstreamRouterName)
+
+	require.Error(t, err)
+}
+
+func TestSetupFailsOnCreateServiceDefaultsForControllerError(t *testing.T) {
+	p, mc := setupPlugin(t)
+
+	testutils.ClearMockCall(&mc.Mock, "CreateServiceDefaults")
+	mc.On("CreateServiceDefaults", mock.Anything).Once().Return(nil)
+	mc.On("CreateServiceDefaults", mock.Anything).Once().Return(fmt.Errorf("boom"))
+
+	err := p.Setup(context.Background())
+
+	mc.AssertCalled(t, "CreateServiceDefaults", "api")
+	mc.AssertCalled(t, "CreateServiceDefaults", clients.ControllerServiceName)
+	mc.AssertNotCalled(t, "CreateServiceDefaults", clients.UpstreamRouterName)
+
+	require.Error(t, err)
+}
+
+func TestSetupFailsOnCreateServiceDefaultsForUpstreamError(t *testing.T) {
+	p, mc := setupPlugin(t)
+
+	testutils.ClearMockCall(&mc.Mock, "CreateServiceDefaults")
+	mc.On("CreateServiceDefaults", mock.Anything).Once().Return(nil)
+	mc.On("CreateServiceDefaults", mock.Anything).Once().Return(nil)
+	mc.On("CreateServiceDefaults", mock.Anything).Once().Return(fmt.Errorf("boom"))
+
+	err := p.Setup(context.Background())
+
+	mc.AssertCalled(t, "CreateServiceDefaults", "api")
+	mc.AssertCalled(t, "CreateServiceDefaults", clients.ControllerServiceName)
+	mc.AssertCalled(t, "CreateServiceDefaults", clients.UpstreamRouterName)
 
 	require.Error(t, err)
 }
@@ -111,14 +152,52 @@ func TestSetupCreatesConsulServiceRouter(t *testing.T) {
 	err := p.Setup(context.Background())
 	require.NoError(t, err)
 
-	mc.AssertCalled(t, "CreateServiceRouter", "api", false)
+	mc.AssertCalled(t, "CreateServiceRouter", "api")
 }
 
 func TestSetupFailsOnCreateServiceRouterError(t *testing.T) {
 	p, mc := setupPlugin(t)
 
 	testutils.ClearMockCall(&mc.Mock, "CreateServiceRouter")
-	mc.On("CreateServiceRouter", mock.Anything, mock.Anything).Return(fmt.Errorf("boom"))
+	mc.On("CreateServiceRouter", mock.Anything).Return(fmt.Errorf("boom"))
+
+	err := p.Setup(context.Background())
+	require.Error(t, err)
+}
+
+func TestSetupCreatesUpstreamServiceRouter(t *testing.T) {
+	p, mc := setupPlugin(t)
+
+	err := p.Setup(context.Background())
+	require.NoError(t, err)
+
+	mc.AssertCalled(t, "CreateUpstreamRouter", "api")
+}
+
+func TestSetupFailsOnCreateUpstreamServiceRouterError(t *testing.T) {
+	p, mc := setupPlugin(t)
+
+	testutils.ClearMockCall(&mc.Mock, "CreateUpstreamRouter")
+	mc.On("CreateUpstreamRouter", mock.Anything).Return(fmt.Errorf("boom"))
+
+	err := p.Setup(context.Background())
+	require.Error(t, err)
+}
+
+func TestSetupCreatesServiceIntention(t *testing.T) {
+	p, mc := setupPlugin(t)
+
+	err := p.Setup(context.Background())
+	require.NoError(t, err)
+
+	mc.AssertCalled(t, "CreateServiceIntention", "api")
+}
+
+func TestSetupFailsOnCreateServiceIntentionError(t *testing.T) {
+	p, mc := setupPlugin(t)
+
+	testutils.ClearMockCall(&mc.Mock, "CreateServiceIntention")
+	mc.On("CreateServiceIntention", mock.Anything).Return(fmt.Errorf("boom"))
 
 	err := p.Setup(context.Background())
 	require.Error(t, err)
@@ -143,13 +222,31 @@ func TestScaleReturnsErrorOnUpdateError(t *testing.T) {
 	require.Error(t, err)
 }
 
+func TestDestroyDeletesServiceSplitter(t *testing.T) {
+	p, mc := setupPlugin(t)
+
+	err := p.Destroy(context.Background())
+	require.NoError(t, err)
+
+	mc.AssertCalled(t, "DeleteServiceSplitter", "api")
+}
+
+func TestDestroyFailsOnDeletesServiceSplitterError(t *testing.T) {
+	p, mc := setupPlugin(t)
+
+	testutils.ClearMockCall(&mc.Mock, "DeleteServiceSplitter")
+	mc.On("DeleteServiceSplitter", mock.Anything).Return(fmt.Errorf("boom"))
+
+	err := p.Destroy(context.Background())
+	require.Error(t, err)
+}
+
 func TestDestroyDeletesServiceRouter(t *testing.T) {
 	p, mc := setupPlugin(t)
 
 	err := p.Destroy(context.Background())
 	require.NoError(t, err)
 
-	mc.AssertCalled(t, "CreateServiceRouter", "api", true)
 	mc.AssertCalled(t, "DeleteServiceRouter", "api")
 }
 
@@ -163,26 +260,26 @@ func TestDestroyFailsOnDeleteServiceRouterError(t *testing.T) {
 	require.Error(t, err)
 }
 
-func TestDestroyDeletesConsulServiceSplitter(t *testing.T) {
+func TestDestroyDeletesUpstreamServiceRouter(t *testing.T) {
 	p, mc := setupPlugin(t)
 
 	err := p.Destroy(context.Background())
 	require.NoError(t, err)
 
-	mc.AssertCalled(t, "DeleteServiceSplitter", "api")
+	mc.AssertCalled(t, "DeleteUpstreamRouter", "api")
 }
 
-func TestDestroyFailsOnDeletesConsulServiceSplitterError(t *testing.T) {
+func TestDestroyFailsOnDeleteUpstreamServiceRouterError(t *testing.T) {
 	p, mc := setupPlugin(t)
 
-	testutils.ClearMockCall(&mc.Mock, "DeleteServiceSplitter")
-	mc.On("DeleteServiceSplitter", mock.Anything).Return(fmt.Errorf("boom"))
+	testutils.ClearMockCall(&mc.Mock, "DeleteUpstreamRouter")
+	mc.On("DeleteUpstreamRouter", mock.Anything).Return(fmt.Errorf("boom"))
 
 	err := p.Destroy(context.Background())
 	require.Error(t, err)
 }
 
-func TestDestroyCreatesConsulServiceResolver(t *testing.T) {
+func TestDestroyDeletesConsulServiceResolver(t *testing.T) {
 	p, mc := setupPlugin(t)
 
 	err := p.Destroy(context.Background())
@@ -191,7 +288,7 @@ func TestDestroyCreatesConsulServiceResolver(t *testing.T) {
 	mc.AssertCalled(t, "DeleteServiceResolver", "api")
 }
 
-func TestDeleteFailsOnCreateServiceResolverError(t *testing.T) {
+func TestDeleteFailsOnDeleteServiceResolverError(t *testing.T) {
 	p, mc := setupPlugin(t)
 
 	testutils.ClearMockCall(&mc.Mock, "DeleteServiceResolver")