+++ plugins/registered_types.go
@@ -1,10 +1,11 @@
 package plugins
 
 const (
-	PluginReleaserTypeConsul    = "consul"
-	PluginRuntimeTypeKubernetes = "kubernetes"
-	PluginMonitorTypePromethus  = "prometheus"
-	PluginStrategyTypeCanary    = "canary"
-	PluginWebhookTypeDiscord    = "discord"
-	PluginWebhookTypeSlack      = "slack"
+	PluginReleaserTypeConsul     = "consul"
+	PluginRuntimeTypeKubernetes  = "kubernetes"
+	PluginMonitorTypePromethus   = "prometheus"
+	PluginStrategyTypeCanary     = "canary"
+	PluginWebhookTypeDiscord     = "discord"
+	PluginWebhookTypeSlack       = "slack"
+	PluginDeploymentTestTypeHTTP = "http"
 )